// src/model/nhanvien.model.ts
export class NhanVien{
    idnhanVien: string=""
    idchiNhanh: string=""
    idtaiKhoan: string=""
    tenNhanVien: string=""
    sdt: string=""
    ngayBatDauLamViec: string=""
    cccd: string=""
    ngaySinh: string=""
    gioiTinh: boolean= true
    email: string=""
    diaChi: string=""
    trangThai: string=""
    ghiChu: string=""
}