// src/model/sanpham.model.ts
export class SanPham{
    idsanPham: string = ""
    idanh: string = ""
    tenSanPham: string = ""
    giaBan: number = 0
    giaVon: number = 0
    donVi: string = ""
    soLuongHienCo: number = 0
    ghiChu: string =""
}

export class HinhAnh{
    
        idanh: string =""
        fileAnh: string = "" 
        ghiChu: string = "" 
        file?: File
      
}